Fix: Fix getSongPosition in AsyncInputManager. & Moved methods.
